+++ clang/test/Analysis/stream-errno-note.c
@@ -1,6 +1,6 @@
 // RUN: %clang_analyze_cc1 -analyzer-checker=core \
 // RUN:   -analyzer-checker=alpha.unix.Stream \
-// RUN:   -analyzer-checker=alpha.unix.Errno \
+// RUN:   -analyzer-checker=unix.Errno \
 // RUN:   -analyzer-checker=unix.StdCLibraryFunctions \
 // RUN:   -analyzer-config unix.StdCLibraryFunctions:ModelPOSIX=true \
 // RUN:   -analyzer-output text -verify %s
@@ -15,7 +15,7 @@ void check_fopen(void) {
   // expected-note@+1{{Taking false branch}}
   if (!F)
     return;
-  if (errno) {} // expected-warning{{An undefined value may be read from 'errno' [alpha.unix.Errno]}}
+  if (errno) {} // expected-warning{{An undefined value may be read from 'errno' [unix.Errno]}}
   // expected-note@-1{{An undefined value may be read from 'errno'}}
   fclose(F);
 }
@@ -27,7 +27,7 @@ void check_tmpfile(void) {
   // expected-note@+1{{Taking false branch}}
   if (!F)
     return;
-  if (errno) {} // expected-warning{{An undefined value may be read from 'errno' [alpha.unix.Errno]}}
+  if (errno) {} // expected-warning{{An undefined value may be read from 'errno' [unix.Errno]}}
   // expected-note@-1{{An undefined value may be read from 'errno'}}
   fclose(F);
 }
@@ -136,7 +136,7 @@ void check_rewind_errnocheck(void) {
     return;
   errno = 0;
   rewind(F); // expected-note{{'rewind' indicates failure only by setting 'errno'}}
-  fclose(F); // expected-warning{{Value of 'errno' was not checked and may be overwritten by function 'fclose' [alpha.unix.Errno]}}
+  fclose(F); // expected-warning{{Value of 'errno' was not checked and may be overwritten by function 'fclose' [unix.Errno]}}
   // expected-note@-1{{Value of 'errno' was not checked and may be overwritten by function 'fclose'}}
 }